requirements.txt
openai>=1.40.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
python-dotenv>=1.0.1
